"en" DataFrame csv files
If they have "en" in the beginning of the name, those are the cleaned DataFrame (to csv) files that are filtered and validated for English detected text. They should have one column for the actual text body and the other for their sentiment label (strings of "positive", "negative", and "neutral"). All NaN values should have been dropped. There should be less than 100k observations for all .csv files